requirements.txt
# Core dependencies for Template Processor
# Template processing engine for TASTE Document Generator
pytest==7.4.2
black==24.3.0
mako==1.3.10
python-docx==1.2.0
bs4==0.0.2
